+++ symboltable.h
@@ -2,13 +2,13 @@ class SymbolTable
 {
 public:
     SymbolTable() {}
-    void insert(string variable, double value);
+    void insert(string variable, int value);
     double lookUp(string variable) const;
     void clear();
 private:
     struct Symbol
     {
-        Symbol(string variable, double value)
+        Symbol(string variable, int value)
         {
             this->variable = variable;
             this->value = value;
